Design: Tasks
Continued work on new scripted tasks. Learned how to create trigger areas with code. Completed two tasks but need to test in multiplayer before I create the rest.

I fixed a bug related to fade to black scene transition. However this led to having to do a rework for the code scene transition which is not finished yet.